# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gemini/conversations:
    get:
      operationId: listGeminiConversations
      tags: [gemini]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeminiConversation"
    post:
      operationId: createGeminiConversation
      tags: [gemini]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGeminiConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiConversation"
  /gemini/conversations/{id}:
    get:
      operationId: getGeminiConversation
      tags: [gemini]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"
    delete:
      operationId: deleteGeminiConversation
      tags: [gemini]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"
  /gemini/conversations/{id}/messages:
    get:
      operationId: listGeminiMessages
      tags: [gemini]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeminiMessage"
    post:
      operationId: sendGeminiMessage
      tags: [gemini]
      summary: Send a message and receive an AI response (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendGeminiMessageBody"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}
  /gemini/generate-image:
    post:
      operationId: generateGeminiImage
      tags: [gemini]
      summary: Generate an image from a text prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateGeminiImageBody"
      responses:
        "200":
          description: Generated image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateGeminiImageResponse"
  /gemini/generate-cover:
    post:
      operationId: generateCover
      tags: [gemini]
      summary: Generate a cover image based on reference image and prompts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateCoverBody"
      responses:
        "200":
          description: Generated cover image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateGeminiImageResponse"
...
  /gemini/improve-cover-idea:
    post:
      operationId: improveCoverIdea
      tags: [gemini]
      summary: Rewrite a rough cover idea into a polished thumbnail title and visual brief
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImproveCoverIdeaBody"
      responses:
        "200":
          description: Improved title and idea text
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImproveCoverIdeaResponse"
  /gemini/generate-youtube-cover:
    post:
      operationId: generateYoutubeCover
      tags: [gemini]
      summary: Generate a horizontal 16:9 YouTube thumbnail (optionally starring a real person from a photo)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateYoutubeCoverBody"
      responses:
        "200":
          description: Generated thumbnail image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateGeminiImageResponse"
  /gemini/adjust-cover:
    post:
      operationId: adjustCover
      tags: [gemini]
      summary: Apply a conversational adjustment to an already generated cover/thumbnail (image edit that keeps everything else intact)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustCoverBody"
      responses:
        "200":
          description: Adjusted image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateGeminiImageResponse"
  /gemini/improve-adjust-instruction:
    post:
      operationId: improveAdjustInstruction
      tags: [gemini]
      summary: Rewrite a rough adjustment request into a clear image-edit instruction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImproveAdjustInstructionBody"
      responses:
        "200":
          description: Rewritten instruction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImproveAdjustInstructionResponse"
...
components:
  schemas:
...
    GeminiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    GeminiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    CreateGeminiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    SendGeminiMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    GeminiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/GeminiMessage"
      required:
        - id
        - title
        - createdAt
        - messages
    GenerateGeminiImageBody:
      type: object
      properties:
        prompt:
          type: string
      required:
        - prompt
    GenerateGeminiImageResponse:
      type: object
      properties:
        b64_json:
          type: string
        mimeType:
          type: string
      required:
        - b64_json
        - mimeType
    GeminiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    GenerateCoverBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        referenceImageBase64:
          type: string
        style:
          type: string
        direccionId:
          type: string
        poseId:
          type: string
        plantillaId:
          type: string
        estiloTitularId:
          type: string
        utileria:
          type: string
      required:
        - title
        - description
    GenerateYoutubeCoverBody:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        style:
          type: string
          maxLength: 300
        personImageBase64:
          type: string
          description: "Legado: una sola foto de la persona (usa personImagesBase64)"
        personImagesBase64:
          type: array
          maxItems: 3
          items:
            type: string
          description: Hasta 3 fotos de LA MISMA persona (más ángulos = más fidelidad de rostro)
        direccionId:
          type: string
        plantillaId:
          type: string
        estiloTitularId:
          type: string
        utileria:
          type: string
          maxLength: 300
      required:
        - title
    AdjustCoverBody:
      type: object
      required: [instruction, imageBase64, formato]
      properties:
        instruction:
          type: string
          maxLength: 1000
        imageBase64:
          type: string
          description: Current image (base64, no data-URL prefix) the adjustment is applied to
        formato:
          type: string
          enum: [vertical, youtube]
    ImproveAdjustInstructionBody:
      type: object
      required: [instruction, formato]
      properties:
        instruction:
          type: string
          maxLength: 1000
        formato:
          type: string
          enum: [vertical, youtube]
        title:
          type: string
          maxLength: 200
    ImproveAdjustInstructionResponse:
      type: object
      required: [instruction]
      properties:
        instruction:
          type: string
    ImproveCoverIdeaBody:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        idea:
          type: string
          maxLength: 2000
        formato:
          type: string
          enum: [vertical, youtube]
        conPersona:
          type: boolean
    ImproveCoverIdeaResponse:
      type: object
      properties:
        title:
          type: string
        idea:
          type: string
        utileria:
          type: string
        estiloExtra:
          type: string
        direccionId:
          type: string
        plantillaId:
          type: string
        estiloTitularId:
          type: string
      required:
        - title
        - idea
...