# lib/api-zod/src/generated/types/generateYoutubeCoverBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateYoutubeCoverBody {
  /** @maxLength 200 */
  title: string;
  /** @maxLength 2000 */
  description?: string;
  /** @maxLength 300 */
  style?: string;
  /** Legado: una sola foto de la persona (usa personImagesBase64) */
  personImageBase64?: string;
  /**
   * Hasta 3 fotos de LA MISMA persona (más ángulos = más fidelidad de rostro)
   * @maxItems 3
   */
  personImagesBase64?: string[];
  direccionId?: string;
  plantillaId?: string;
  estiloTitularId?: string;
  /** @maxLength 300 */
  utileria?: string;
}